template/requirements/_tests.txt
git+https://github.com/aloetesting/aloe.git@e8392800bce42faffd5e94ea358dbbb69b7989fa#egg=aloe
git+https://github.com/jricardo27/aloe_django@c757f8fe27b37756482f4de7ee246284c294e550#egg=aloe_django
aloe_webdriver==0.5.6
coverage==4.4.1
django_coverage_plugin==1.5.0
factory_boy==2.8.1
pytest==3.1.3
pytest-cov==2.5.1
pytest-django==3.1.2
selenium==3.4.3
xvfbwrapper==0.2.9
